Fix TS watcher issue in two step compilation
