+++ lib/TypescriptCompiler.js
@@ -26,13 +26,14 @@
 
 "use strict";
 
-let fs = require('fs-extra');
-let path = require('path');
-let ts = require('typescript');
-let _ = require('lodash');
-let concat = require('source-map-concat');
-let config = require('../config.json');
-let ConsoleWriter = require('./ConsoleWriter');
+const fs = require('fs-extra');
+const path = require('path');
+const ts = require('typescript');
+const _ = require('lodash');
+const concat = require('source-map-concat');
+const config = require('../config.json');
+const ConsoleWriter = require('./ConsoleWriter');
+const encoding = "utf8";
 
 const PLUGIN_TEMPLATE_PATH = path.join(__dirname, '..', config.templates.plugin);
 
@@ -43,7 +44,7 @@ const PLUGIN_TEMPLATE_PATH = path.join(__dirname, '..', config.templates.plugin)
  * @param {object} options - TSC compiler options object
  * @returns {Promise}
  */
-function runWatcher(rootFileNames, options, packagemode) {
+function runWatcher(rootFileNames, options, packagemode, visualPackage) {
     const files = {};
     for (let fileName in rootFileNames) {
         files[rootFileNames[fileName]] = { version: 0 };
@@ -106,6 +107,8 @@ function runWatcher(rootFileNames, options, packagemode) {
 
                     // write the changes to disk
                     emitFile(fileName);
+                    concatExternalJS(visualPackage);
+                    appendExportPowerBINameSpace(visualPackage, options);
                 });
         });
     }
@@ -196,8 +199,24 @@ async function concatExternalJS(visualPackage) {
     await fs.writeFile(visualJsMapPath, concatResult.map.toString());
 }
 
+async function appendExportPowerBINameSpace(visualPackage, compilerOptions) {
+    if (compilerOptions.outDir) {
+        // we should not use externalJS in modern style modules
+        return;
+    }
+    let visualJsName = compilerOptions.out;
+    let visualJSFileContent = "";
+    let visualJSFilePath = visualPackage.buildPath(visualJsName);
+
+    visualJSFileContent += "\n" + fs.readFileSync(visualJSFilePath, { encoding: encoding });
+    visualJSFileContent = "\nvar powerbi = globalPowerbi;\n" + visualJSFileContent;
+    visualJSFileContent += "\nmodule.exports = { powerbi };";
+    await fs.writeFile(visualJSFilePath, visualJSFileContent);
+}
+
 module.exports = {
     concatExternalJS,
     runWatcher,
-    createPlugin
+    createPlugin,
+    appendExportPowerBINameSpace
 };
