+++ lib/WebPackWrap.js
@@ -15,6 +15,7 @@ const uglifyJS = require('uglify-js');
 const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
 const util = require('util');
 const exec = util.promisify(require('child_process').exec);
+const _ = require("lodash");
 
 class WebPackGenerator {
 
@@ -31,21 +32,6 @@ class WebPackGenerator {
         if (!fs.existsSync(dropFolder)) {
             fs.mkdirSync(dropFolder);
         }
-    }
-
-    static async appendExportPowerBINameSpace(visualPackage, tsconfig) {
-        if (tsconfig.compilerOptions.outDir) {
-            // we should not use externalJS in modern style modules
-            return;
-        }
-        let visualJsName = tsconfig.compilerOptions.out;
-        let visualJSFileContent = "";
-        let visualJSFilePath = visualPackage.buildPath(visualJsName);
-
-        visualJSFileContent += "\n" + fs.readFileSync(visualJSFilePath, { encoding: encoding });
-        visualJSFileContent = "\nvar powerbi = globalPowerbi;\n" + visualJSFileContent;
-        visualJSFileContent += "\nmodule.exports = { powerbi };";
-        await fs.writeFile(visualJSFilePath, visualJSFileContent);
     }
 
     static loadAPIPackage() {
@@ -66,7 +52,7 @@ class WebPackGenerator {
         const visualJsOutDir = tsconfig.compilerOptions.outDir;
         const visualJSFilePath = visualPackage.buildPath(visualJsName || visualJsOutDir);
 
-        let pluginConfiguration = visualPackage.config;
+        let pluginConfiguration = _.cloneDeep(visualPackage.config);
 
         if (tsconfig.compilerOptions.outDir) {
             let api = this.loadAPIPackage(visualPackage);
@@ -107,7 +93,7 @@ class WebPackGenerator {
         pluginConfiguration.generateResources = options.generateResources;
         pluginConfiguration.minifyJS = options.minifyJS;
         pluginConfiguration.dependencies = pbiviz.dependencies;
-        pluginConfiguration.modules = tsconfig.compilerOptions.outDir !== null;
+        pluginConfiguration.modules = typeof tsconfig.compilerOptions.outDir !== "undefined";
         pluginConfiguration.visualSourceLocation = path.posix.relative(config.build.precompileFolder, tsconfig.files[0]).replace(".ts", "");
 
         if (options.minifyJS) {
@@ -157,20 +143,18 @@ class WebPackGenerator {
             passphrase: config.server.passphrase
         };
 
-        const pluginFileName = "visualPlugin.ts";
-        let pluginLocation = path.join(process.cwd(), ".tmp", "precompile", pluginFileName);
-        webpackConfig.plugins.push(
-            new webpack.WatchIgnorePlugin([pluginLocation])
-        );
         if (tsconfig.compilerOptions.out) {
             webpackConfig.entry = {
                 "visual.js": [visualJSFilePath]
             };
         }
         else {
-            webpackConfig.entry  = {
-                "visual.js": [pluginLocation]
-            }
+            const pluginFileName = "visualPlugin.ts";
+            let pluginLocation = path.join(process.cwd(), ".tmp", "precompile", pluginFileName);
+            webpackConfig.entry["visual.js"] = [pluginLocation];
+            webpackConfig.plugins.push(
+                new webpack.WatchIgnorePlugin([pluginLocation])
+            );
         }
 
         return webpackConfig;
@@ -242,11 +226,11 @@ class WebPackGenerator {
                 );
             tsconfig.files.push(pluginDropPath);
 
-            await TypescriptCompiler.runWatcher(tsconfig.files, tsconfig.compilerOptions, !options.devMode, pbiviz.visual.guid);
+            await TypescriptCompiler.runWatcher(tsconfig.files, tsconfig.compilerOptions, !options.devMode, visualPackage);
             await TypescriptCompiler.concatExternalJS(visualPackage);
             // await WebPackGenerator.assemblyExternalJSFiles(visualPackage, options.minifyJS, tsconfig.compilerOptions.out);
             await LessCompiler.build(visualPackage, options);
-            await WebPackGenerator.appendExportPowerBINameSpace(visualPackage, tsconfig);
+            await TypescriptCompiler.appendExportPowerBINameSpace(visualPackage, tsconfig.compilerOptions);
             webpackConfig = await WebPackGenerator.prepareWebPackConfig(visualPackage, options, tsconfig);
         }
         return webpackConfig;
